Update IntegerCalculations.cs
Change the return value of Product method from int to long.

+++ Homeworks/03. CSharp-Advance/03. Methods/IntegerCalculations/IntegerCalculations.cs
@@ -46,9 +46,9 @@ class IntegerCalculations
         return result;
     }
 
-    static int Product(int[] array)
+    static long Product(int[] array)
     {
-        int result = 1;
+        long result = 1;
         for (int i = 0; i < array.Length; i++)
         {
             result = result * array[i];
